+++ mne/viz/evoked.py
@@ -11,6 +11,7 @@ from __future__ import print_function
 # License: Simplified BSD
 
 from functools import partial
+from copy import deepcopy
 
 import numpy as np
 
@@ -19,7 +20,7 @@ from ..io.pick import (channel_type, pick_types, _picks_by_type,
 from ..externals.six import string_types
 from ..defaults import _handle_default
 from .utils import (_draw_proj_checkbox, tight_layout, _check_delayed_ssp,
-                    plt_show, _process_times, DraggableColorbar)
+                    plt_show, _process_times, DraggableColorbar, _setup_cmap)
 from ..utils import logger, _clean_names, warn
 from ..io.pick import pick_info
 from .topo import _plot_evoked_topo
@@ -189,10 +190,8 @@ def _plot_evoked(evoked, picks, exclude, unit, show,
                            ' for interactive SSP selection.')
     if isinstance(gfp, string_types) and gfp != 'only':
         raise ValueError('gfp must be boolean or "only". Got %s' % gfp)
-    if cmap == 'interactive':
-        cmap = (None, True)
-    elif not isinstance(cmap, tuple):
-        cmap = (cmap, True)
+    cmap = _setup_cmap(cmap)
+
     scalings = _handle_default('scalings', scalings)
     titles = _handle_default('titles', titles)
     units = _handle_default('units', units)
@@ -1500,7 +1499,12 @@ def plot_compare_evokeds(evokeds, picks=list(), gfp=False, colors=None,
     # are pulled from the 'colors' and 'linestyles' dicts via '/'-tag matching
     # unless they are overwritten by entries from a user-provided 'styles'.
 
-    # first, check if input is valid
+    # first, copy to avoid overwriting
+    styles = deepcopy(styles)
+    colors = deepcopy(colors)
+    linestyles = deepcopy(linestyles)
+
+    # second, check if input is valid
     if isinstance(styles, dict):
         for style_ in styles:
             if style_ not in conditions:
@@ -1528,14 +1532,14 @@ def plot_compare_evokeds(evokeds, picks=list(), gfp=False, colors=None,
     else:
         colors = _setup_styles(conditions, colors, "color", "grey")
 
-    # third, linestyles
+    # fourth, linestyles
     if not isinstance(linestyles, dict):
         linestyles = dict((condition, linestyle) for condition, linestyle in
                           zip(conditions, ['-'] * len(conditions)))
     else:
         linestyles = _setup_styles(conditions, linestyles, "linestyle", "-")
 
-    # fourth, put it all together
+    # fifth, put it all together
     if styles is None:
         styles = dict()
     for condition, color, linestyle in zip(conditions, colors, linestyles):